Update custom user model to overwrite Allauth Signup
Remove username as a requirement for registering an account.

Add first name, last name, date of birth and phone number as a requirement for registering both users and superusers.

Update the admin to register the updated user model.

Add configuration settings to make sure the custom user model works with Allauth.

Django extensions was used to reset the database via the command line.